requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
sqlalchemy==2.0.20
psycopg2-binary==2.9.7  
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
aiohttp==3.8.5
jinja2==3.1.2
python-dotenv==1.0.0
alembic==1.12.0
pydantic-settings==2.0.3  